+++ frontend/src/App.css
@@ -1,42 +1,98 @@
-#root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
+/* Reset any default list styling */
+ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
 }
 
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+/* Navigation Bar */
+.navbar {
+  position: fixed; /* Change to fixed */
+  top: 0;
+  width: 100%; /* Ensure it spans the full width */
+  z-index: 1000; /* Ensures it stays above other content */
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background-color: #1E90FF; /* Nice blue color */
+  padding: 1rem 2rem;
+  color: #fff;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
+.navbar-logo {
+  display: flex;
+  align-items: center;
+  font-size: 1.5rem;
+  font-weight: bold;
 }
 
-.card {
-  padding: 2em;
+.navbar-logo img {
+  width: 50px; /* Adjusted to 50px */
+  height: 50px; /* Adjusted to 50px */
+  margin-right: 0.5rem;
+  margin-top: 0.4rem;
 }
 
-.read-the-docs {
-  color: #888;
+.navbar-links {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.nav-item {
+  font-size: 1.1rem;
+}
+
+.nav-item button,
+.nav-item a {
+  background: none;
+  border: none;
+  color: #fff;
+  text-decoration: none;
+  cursor: pointer;
+  font-size: inherit;
+  padding: 0.5rem 1rem;
+  transition: background-color 0.2s ease-in-out;
+}
+
+.nav-item button:hover,
+.nav-item a:hover {
+  background-color: rgba(255, 255, 255, 0.2);
+}
+
+/* Profile Section */
+.profile-container {
+  position: relative;
+  display: inline-block;
+}
+
+.profile-image {
+  width: 50px; /* Adjusted to 50px */
+  height: 50px; /* Adjusted to 50px */
+  border-radius: 50%;
+  object-fit: cover;
+  display: block;
+  margin-right: 2.5em;
+}
+
+/* Tooltip-style name on hover */
+.profile-name {
+  position: absolute;
+  top: 80px;  /* Position below the image */
+  right: -15%;
+  transform: translateX(-50%);
+  background-color: #1E90FF;
+  color: #fff;
+  font-size: 0.8rem;
+  padding: 0.25rem 0.5rem;
+  border-radius: 0.3rem;
+  white-space: nowrap;
+  opacity: 0;
+  visibility: hidden;
+  transition: opacity 0.2s ease-in-out;
+}
+
+.profile-container:hover .profile-name {
+  opacity: 1;
+  visibility: visible;
 }
