+++ frontend/src/components/EntraLogon.jsx
@@ -1,9 +1,8 @@
-
 import React from 'react';
-import { useMsal , AuthenticatedTemplate, UnauthenticatedTemplate} from '@azure/msal-react';
-import { loginRequest } from './components/azureAuth';
+import { useMsal, AuthenticatedTemplate, UnauthenticatedTemplate } from '@azure/msal-react';
+import { loginRequest } from './azureAuth';
 
-const AzureLogon = () => {
+const EntraLogon = () => {
   const { instance } = useMsal();
 
   const logonFunc = async (forcePopup = false) => {
@@ -11,28 +10,28 @@ const AzureLogon = () => {
       let loginRequestParam = forcePopup ? { ...loginRequest, prompt: "select_account" } : loginRequest;
       const response = await instance.loginPopup(loginRequestParam);
       instance.setActiveAccount(response.account);
-
     } catch (error) {
       console.error("Login failed:", error);
     }
   };
 
-
   const logoutFunc = async () => {
     await instance.logoutRedirect({
       postLogoutRedirectUri: window.location.origin // or any URL you want users redirected to
     });
-  }
+  };
 
-  return <>
+  return (
+    <div>
       <AuthenticatedTemplate>
-        <button onClick={logoutFunc}>Sign Out from Azure</button>
-        <button onClick={()=>logonFunc(true)}>Sign In with Different Account</button>
+        <a href="#" onClick={logoutFunc} className="nav-link">Sign Out</a>
+        <a href="#" onClick={() => logonFunc(true)} className="nav-link">Change Account</a>
       </AuthenticatedTemplate>
       <UnauthenticatedTemplate>
-        <button onClick={logoutFunc}>Sign In with Azure</button>
+        <a href="#" onClick={() => logonFunc()} className="nav-link">Sign In</a>
       </UnauthenticatedTemplate>
-    </>
+    </div>
+  );
 };
 
-export default AzureLogon;
+export default EntraLogon;
