+++ frontend/src/components/azureAuth.js
@@ -32,38 +32,5 @@ export const loginRequest = {
   scopes: ['User.Read'],
 };
 
-// Custom function to receive the profile phote
-export const getProfilePhoto = async (instance, activeAccount) => {
-    try {
-      if (!activeAccount && !instance) {
-        console.warn('No active account or instance');
-        return;
-      }
-      // Request token for Microsoft Graph with the User.Read scope
-      const tokenResponse = await instance.acquireTokenSilent({
-        ...loginRequest,
-        account: activeAccount,
-      });
-      
-      const accessToken = tokenResponse.accessToken;
-      
-      // Fetch the profile photo from Graph
-      const response = await fetch('https://graph.microsoft.com/v1.0/me/photo/$value', {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      });
-      
-      if (response.ok) {
-        const blob = await response.blob();
-        return URL.createObjectURL(blob);
-        
-      } else {
-        console.error('Failed to fetch profile photo:', response.statusText);
-      }
-    } catch (error) {
-      if (error.errorMessage !== "You must call and await the initialize function before attempting to call any other MSAL API.  For more visit: aka.ms/msaljs/browser-errors")
-        console.error('Error fetching profile photo:', error);
-    }
-};
+
